+++ test/integration/verify/controls/aws_elasticache_cluster.rb
@@ -7,12 +7,12 @@ aws_elasticache_cluster_parameter_group_name = attribute(:aws_elasticache_cluste
 
 title 'Test single AWS ElastiCache Cluster'
 
-control 'aws-elasti-cache-cluster-1.0' do
+control 'aws-elasticache-cluster-1.0' do
 
   impact 1.0
   title 'Ensure AWS ElastiCache Cluster has the correct properties.'
 
-  describe aws_elasti_cache_cluster(cache_cluster_id: aws_elasticache_cluster_id) do
+  describe aws_elasticache_cluster(cache_cluster_id: aws_elasticache_cluster_id) do
     it { should exist }
     it { should_not be_encrypted_at_rest }
     it { should_not be_encrypted_at_transit }
@@ -23,7 +23,7 @@ control 'aws-elasti-cache-cluster-1.0' do
     its('cache_parameter_group.cache_parameter_group_name') { should eq aws_elasticache_cluster_parameter_group_name }
   end
 
-  describe aws_elasti_cache_cluster('not-there') do
+  describe aws_elasticache_cluster('not-there') do
     it { should_not exist }
   end
 end
