+++ test/unit/resources/aws_elasticache_clusters_test.rb
@@ -1,7 +1,7 @@
 require 'helper'
-require 'aws_elasti_cache_clusters'
+require 'aws_elasticache_clusters'
 require "aws-sdk-core"
-require_relative 'mock/aws_elasti_cache_cluster_mock'
+require_relative 'mock/aws_elasticache_cluster_mock'
 
 class AwsElastiCacheClustersConstructorTest < Minitest::Test
