+++ test/unit/resources/mock/aws_elasticache_cluster_mock.rb
